Add config setting for vapor (#2682)
* Add test

* Add temporary file deletion to read chunk

* Use config to prevent deletion of non-existant local file

* Add config setting

* Format code

* Fix test fake

+++ src/Files/RemoteTemporaryFile.php
@@ -80,7 +80,10 @@ class RemoteTemporaryFile extends TemporaryFile
      */
     public function delete(): bool
     {
-        $this->deleteLocalCopy();
+        // we don't need to delete local copy as it's deleted at end of each chunk
+        if (!config('excel.temporary_files.force_resync_remote')) {
+            $this->deleteLocalCopy();
+        }
 
         return $this->disk()->delete($this->filename);
     }
